+++ docs/changes.rst
@@ -4,7 +4,7 @@ Changelog
 14.1.3 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Users can enter and edit their first name and last name.
 
 
 14.1.2 (2022-06-15)